detect initial working directory
In order to correctly resolve paths with realpath() when the current
working directory has changed from what it was initially, we need to
record what it was initially.  As there are limited means for doing
this in C (compiler extensions), we leverage features of C++ that
ensure code is run when the library is loaded (typically before
main()).  There are some alternative methods at disposal such as
inspecting /proc on some platforms, but testing first if this
pragmatic forced detection approach is reliable.