progipwd() implementation for getting initial working directory
In order for progpath() to make reliable use of program path methods
that return a relative filesystem path, we need to know what the
original working directory was when the program was launched.  We
intentionally avoid the typical approach having apps get+save the path
themselves, requiring some function be called early on, as we don't
want that call to be necessary in order for progpath() to work.

What we do instead is treat it like progpath() and utilize various
platform-specific methods at our disposal.  There are very limited
options and many platforms that simply do not record the path, so a
method for automatically querying and saving the path was devised
utilizing C++ object initialization.  As this is independently useful
from identifying the program path, we keep it as a second API call.

If we're going to rely on C++, it may make more sense now to turn this
into header-only.